backend/requirements.txt
fastapi
uvicorn
qdrant-client
openai
apscheduler
pydantic
pydantic-settings
python-dotenv
langdetect
